refactor(clip): move the clip render kernel out of AudioEngine
Pure relocation. renderClipInto becomes a free function in
Core/ClipRenderKernel, so the consolidation renderer can call the same
code the live path calls rather than re-deriving an interpretation of
ClipEdits — the requirement in consolidate-audio-range.md §3.1.

Unlike the previous commit, the six unchanged digests are real evidence
here. Relocation is exactly what those fixtures can observe, whereas the
quality normalization was invisible to them by construction.

The kernel holds no engine state: no transport, graph, routing, PDC or
automation, and after the previous commit no atomic either. Quality
arrives as a value because the engine snapshots it once per block, which
is why the normalization had to land first — a function outside
AudioEngine has nothing to load it from. AudioEngine.h no longer
declares the helper at all.

MixMath.h returns here, where it is actually needed: a separate
translation unit cannot see the file-local clampD and sanitizeMix, and
copying them would seed the divergence the contract forbids. It was
dropped from B1 because a private member in the same TU reached them
directly.

Note, not fixed here: CLIP_EDGE_FADE_SAMPLES was already defined twice,
in AudioEngine.h and AudioRenderer.h. The kernel adds
kClipEdgeFadeSamples as its own authority, making three. Consolidating
them is a separate change rather than something to fold into a move.

Audio baseline 6/6 against develop@6b79af40, SRC telemetry 29/29,
canonical suite 224/224.
